Add a simple AssemblyCatalogVersionSource based on mtime and length.
With this, the cacher should be usable at an assembly granularity
(i.e., an instance of CachingCatalog, AssemblyCatalogVersionSource,
SerializingStoringPartCache, FileCacheStorage, and
SimplePartSerializer for each assembly). Probably not optimally
performant, though.

AssemblyCatalogVersionSource requires the filename of the assembly be
specified. It does not try to infer information about the catalog
passed to its GetVersion() function. In fact, Iâ€™m not sure how things
would be able to make use of that parameter (without resorting to
upcasting). Might need to rethink that.